+++ tools/sdk-sync/src/lib.rs
@@ -1,0 +1,10 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+pub mod fs;
+pub mod git;
+pub mod gradle;
+pub mod sync;
+pub mod versions;
